requirements.txt
torch>=2.0
torchvision>=0.15
torchaudio
scikit-image
scikit-learn
matplotlib
pandas
numpy
pillow
seaborn
tqdm
ipywidgets
notebook


# enable GPU-capable PyTorch wheels (includes CUDA runtime)
--extra-index-url https://download.pytorch.org/whl/cu126